NOTIF-150 Introduce behavior groups - Phase 2
