+++ src/main/java/com/redhat/cloud/notifications/events/EndpointProcessor.java
@@ -3,7 +3,6 @@ package com.redhat.cloud.notifications.events;
 import com.redhat.cloud.notifications.db.EndpointResources;
 import com.redhat.cloud.notifications.db.NotificationResources;
 import com.redhat.cloud.notifications.ingress.Action;
-import com.redhat.cloud.notifications.models.Endpoint;
 import com.redhat.cloud.notifications.models.EndpointType;
 import com.redhat.cloud.notifications.models.Notification;
 import com.redhat.cloud.notifications.processors.EndpointTypeProcessor;
@@ -12,13 +11,11 @@ import com.redhat.cloud.notifications.processors.email.EmailSubscriptionTypeProc
 import com.redhat.cloud.notifications.processors.webhooks.WebhookTypeProcessor;
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.MeterRegistry;
-import io.smallrye.mutiny.Multi;
 import io.smallrye.mutiny.Uni;
 
 import javax.annotation.PostConstruct;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
-import java.util.function.Function;
 import java.util.logging.Logger;
 
 @ApplicationScoped
@@ -31,10 +28,6 @@ public class EndpointProcessor {
 
     @Inject
     EndpointResources resources;
-
-    // TODO [BG Phase 2] Delete this attribute
-    @Inject
-    DefaultProcessor defaultProcessor;
 
     @Inject
     NotificationResources notifResources;
@@ -62,8 +55,7 @@ public class EndpointProcessor {
 
     public Uni<Void> process(Action action) {
         processedItems.increment();
-        // TODO [BG Phase 2] Use EndpointResources.getTargetEndpoints here
-        return getEndpoints(
+        return resources.getTargetEndpoints(
                 action.getAccountId(),
                 action.getBundle(),
                 action.getApplication(),
@@ -90,17 +82,5 @@ public class EndpointProcessor {
             default:
                 return notification -> Uni.createFrom().nullItem();
         }
-    }
-
-    // TODO [BG Phase 2] Delete this method
-    public Multi<Endpoint> getEndpoints(String tenant, String bundleName, String applicationName, String eventTypeName) {
-        return resources.getTargetEndpoints(tenant, bundleName, applicationName, eventTypeName)
-                .onItem().transformToMultiAndConcatenate((Function<Endpoint, Multi<Endpoint>>) endpoint -> {
-                    // If the tenant has a default endpoint for the eventType, then add the target endpoints here
-                    if (endpoint.getType() == EndpointType.DEFAULT) {
-                        return defaultProcessor.getDefaultEndpoints(endpoint);
-                    }
-                    return Multi.createFrom().item(endpoint);
-                });
     }
 }
