+++ src/main/java/com/redhat/cloud/notifications/models/EndpointType.java
@@ -12,7 +12,14 @@ public enum EndpointType {
     WEBHOOK, // 0
     @JsonProperty("email_subscription")
     EMAIL_SUBSCRIPTION, // 1
-    @JsonProperty("default") // TODO [BG Phase 2] Delete the DEFAULT enum member
+    /**
+     * This enum member is no longer used.
+     * <p>
+     * It <b>MUST NOT</b> be deleted because the {@link EndpointType} members ordinal is stored into the database.
+     * @deprecated Since the behavior groups migration.
+     */
+    @JsonProperty("default")
+    @Deprecated
     DEFAULT, // 2
     @JsonProperty("camel")
     CAMEL // 3
