+++ src/main/resources/application.properties
@@ -13,6 +13,7 @@ mp.messaging.incoming.ingress.topic=platform.notifications.ingress
 mp.messaging.incoming.ingress.group.id=integrations
 mp.messaging.incoming.ingress.key.deserializer=org.apache.kafka.common.serialization.StringDeserializer
 mp.messaging.incoming.ingress.value.deserializer=org.apache.kafka.common.serialization.StringDeserializer
+%prod.mp.messaging.incoming.ingress.enabled=false # TODO [BG Phase 2.1] Remove this line
 
 # Output queue to Camel (notifications-sender)
 mp.messaging.outgoing.toCamel.connector=smallrye-kafka
