+++ src/test/java/com/redhat/cloud/notifications/db/ResourceHelpers.java
@@ -2,6 +2,7 @@ package com.redhat.cloud.notifications.db;
 
 import com.redhat.cloud.notifications.TestHelpers;
 import com.redhat.cloud.notifications.models.Application;
+import com.redhat.cloud.notifications.models.BehaviorGroup;
 import com.redhat.cloud.notifications.models.Bundle;
 import com.redhat.cloud.notifications.models.EmailAggregation;
 import com.redhat.cloud.notifications.models.EmailSubscription;
@@ -15,6 +16,7 @@ import com.redhat.cloud.notifications.models.WebhookProperties;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 @ApplicationScoped
@@ -40,6 +42,9 @@ public class ResourceHelpers {
     @Inject
     EmailAggregationResources emailAggregationResources;
 
+    @Inject
+    BehaviorGroupResources behaviorGroupResources;
+
     public List<Application> getApplications(String bundleName) {
         return appResources.getApplications(bundleName).await().indefinitely();
     }
@@ -48,7 +53,7 @@ public class ResourceHelpers {
         return subscriptionResources.getEmailSubscription(accountNumber, username, bundle, application, type).await().indefinitely();
     }
 
-    public void createTestAppAndEventTypes() {
+    public UUID createTestAppAndEventTypes() {
         Bundle bundle = new Bundle(TEST_BUNDLE_NAME, "...");
         Bundle b = bundleResources.createBundle(bundle).await().indefinitely();
 
@@ -82,6 +87,8 @@ public class ResourceHelpers {
             eventType.setDisplayName("... -> " + i);
             appResources.createEventType(eventType).await().indefinitely();
         }
+
+        return b.getId();
     }
 
     public int[] createTestEndpoints(String tenant, int count) {
@@ -94,13 +101,8 @@ public class ResourceHelpers {
             properties.setUrl("https://localhost");
 
             Endpoint ep = new Endpoint();
-            if (i > 0) {
-                ep.setType(EndpointType.WEBHOOK);
-                ep.setName(String.format("Endpoint %d", count - i));
-            } else {
-                ep.setType(EndpointType.DEFAULT);
-                ep.setName("Default endpoint");
-            }
+            ep.setType(EndpointType.WEBHOOK);
+            ep.setName(String.format("Endpoint %d", count - i));
             ep.setDescription("Automatically generated");
             boolean enabled = (i % (count / 5)) != 0;
             if (!enabled) {
@@ -132,24 +134,6 @@ public class ResourceHelpers {
         return resources.createEndpoint(ep).await().indefinitely().getId();
     }
 
-    public void assignEndpointToEventType(String tenant, UUID endpointId, UUID eventTypeId) {
-        resources.linkEndpoint(tenant, endpointId, eventTypeId).await().indefinitely();
-    }
-
-    public UUID getDefaultEndpointId(String tenant) {
-        Endpoint ep = new Endpoint();
-        ep.setType(EndpointType.DEFAULT);
-        ep.setName(String.format("Endpoint %s", UUID.randomUUID().toString()));
-        ep.setDescription("Automatically generated");
-        ep.setEnabled(true);
-        ep.setAccountId(tenant);
-
-        return resources.createEndpoint(ep).await().indefinitely().getId();
-    }
-
-    public void assignEndpointToDefault(String tenant, UUID endpointId) {
-        resources.addEndpointToDefaults(tenant, endpointId).await().indefinitely();
-    }
 
     public void createSubscription(String tenant, String username, String bundle, String application, EmailSubscriptionType type) {
         subscriptionResources.subscribe(tenant, username, bundle, application, type).await().indefinitely();
@@ -166,5 +150,20 @@ public class ResourceHelpers {
 
     public List<EventType> getEventTypesForApplication(UUID applicationId) {
         return appResources.getEventTypes(applicationId).await().indefinitely();
+    }
+
+    public UUID createBehaviorGroup(String accountId, String name, UUID bundleId) {
+        BehaviorGroup behaviorGroup = new BehaviorGroup();
+        behaviorGroup.setDisplayName("Behavior group");
+        behaviorGroup.setBundleId(bundleId);
+        return behaviorGroupResources.create(accountId, behaviorGroup).await().indefinitely().getId();
+    }
+
+    public void updateEventTypeBehaviors(String accountId, UUID eventTypeId, Set<UUID> behaviorGroupIds) {
+        behaviorGroupResources.updateEventTypeBehaviors(accountId, eventTypeId, behaviorGroupIds).await().indefinitely();
+    }
+
+    public void updateBehaviorGroupActions(String accountId, UUID behaviorGroupId, List<UUID> endpointIds) {
+        behaviorGroupResources.updateBehaviorGroupActions(accountId, behaviorGroupId, endpointIds).await().indefinitely();
     }
 }
